+++ src/components/BubbleChat/index.js
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import PropTypes, { any } from 'prop-types';
 import classNames from 'classnames';
-import Avatar from '../Avatar';
+import Avatar from 'components/Avatar';
 import BubbleChatImage from './Image';
 import Context from './Context';
 
@@ -152,6 +152,7 @@ const BubbleChat = React.forwardRef(({ className, isTyping, text, type, variant,
                     'BubbleChat-typing',
                     'u-paddingExtraSmall u-positionRelative',
                     type && typeThemeClassNames[type],
+                    variantOri && variantTextClassNames[variantOri],
                     variantOri && variantClassNames[variantOri],
                   )}
                 >