+++ src/components/Icon/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { icons } from '../../constants';
+import { icons } from 'constants/icons';
 
 const propTypes = {
   /** The icon visual name */
@@ -145,6 +145,21 @@ const propTypes = {
     'objects',
     'reply',
     'bot',
+    'shapes',
+    'return',
+    'umbrella',
+    'game',
+    'tagCloud',
+    'one',
+    'two',
+    'three',
+    'four',
+    'five',
+    'six',
+    'seven',
+    'eight',
+    'nine',
+    'ten',
   ]),
   /** Icon size variants */
   size: PropTypes.oneOf([